+++ docs/changelog.md
@@ -39,6 +39,7 @@ This changelog follows the rules of [Keep a Changelog](http://keepachangelog.com
 - Support for **navigating through the menu using the arrow keys**. Highlight items with the arrow keys and press <kbd>Enter</kbd> to select them.
 - A **redesigned General-Settings dialog** with a completely new layout. All the options are still there, but they are now organized in categories which should make it easier to navigate through the settings.
 - New options in the menu-themes dialog to **disable the animation of menu items**. Thanks to [@Yavuz-Kagan-Yadigar](https://github.com/Yavuz-Kagan-Yadigar) for contributing this feature!
+- A global option to **return the pointer to the menu opening position** after selecting a button.
 
 ### :wrench: Changed